requirements.txt
langchain==0.0.184
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
youtube-transcript-api==0.6.1
beautifulsoup4=4.12.2
pytube==15.0.0
chromadb==0.3.26
deepgram-sdk==2.12.0
pyaudio==0.2.14
rhasspy-silence==0.4.0
elevenlabs==0.2.27
pygame==2.5.2
pyaudio